Fix audit items: tax card type error, migration collision, buy button, feed polish
1. Fix .clamp() returning num → added .toDouble() casts on tax card
2. Rename migration 20260211000001_scheduled_followup → 000002 (timestamp collision)
3. Product buy button now opens WhatsApp to arrange purchase (was empty)
4. Product showcase card in feed: dedicated layout with image, brand, price, CTA

+++ beautycita_app/lib/screens/business/business_dashboard_screen.dart
@@ -430,12 +430,12 @@ class _TaxDeductionsCardState extends ConsumerState<_TaxDeductionsCard> {
     final estimatedIvaTotal = _revenueYtd * 0.16;
     final estimatedIsrTotal = _revenueYtd * 0.10; // simplified estimate
     final estimatedTotalTax = estimatedIvaTotal + estimatedIsrTotal;
-    final taxStillOwed = (estimatedTotalTax - totalTaxPaid).clamp(0, double.infinity);
+    final taxStillOwed = (estimatedTotalTax - totalTaxPaid).clamp(0.0, double.infinity).toDouble();
 
     // Deduction budget: expenses reduce taxable income
     // The salon can deduct 100% of business expenses
     // Show how much more they could spend to offset remaining tax
-    final deductionBudget = (_revenueYtd - _expensesYtd).clamp(0, double.infinity);
+    final deductionBudget = (_revenueYtd - _expensesYtd).clamp(0.0, double.infinity).toDouble();
 
     return Card(
       elevation: 0,
